refactor: extract common item functions
