+++ compare-craft.html
@@ -97,6 +97,7 @@
   <script src="js/item-tabs.js"></script>
   <script src="js/feedback-modal.js"></script>
   <!-- Nuevos módulos separados -->
+  <script type="module" src="js/items-core.js"></script>
   <script type="module" src="js/compare-logic.js"></script>
   <script type="module" src="js/compare-ui.js"></script>
