+++ item.html
@@ -130,6 +130,7 @@
 <script src="js/itemHandlers.js"></script>
 <script src="js/storageUtils.js"></script>
 <script src="js/item-ui.js"></script>
+<script type="module" src="js/items-core.js"></script>
 <script type="module" src="js/item.js"></script>
   
   <!-- Inicialización -->
